# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NFL Survivor Pool API",
    "version": "1.0.0",
    "description": "API for NFL Survivor Pool Web App"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Pick": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "entry": {
            "type": "string"
          },
          "entryNumber": {
            "type": "integer"
          },
          "team": {
            "type": "string"
          },
          "week": {
            "type": "integer"
          },
          "game": {
            "type": "object",
            "properties": {
              "_id": {
                "type": "string"
              },
              "away_team": {
                "type": "string"
              },
              "home_team": {
                "type": "string"
              },
              "event_date": {
                "type": "string",
                "format": "date-time"
              },
              "schedule": {
                "type": "object",
                "properties": {
                  "week": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Unauthorized access",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "NotFoundError": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "BadRequestError": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/picks/pool/{poolId}": {
      "get": {
        "summary": "Get all picks for a pool",
        "tags": [
          "Picks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "poolId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the pool"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Pick"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
    "/api/v1/picks/{entryId}/{entryNumber}/{week}": {
      "get": {
        "summary": "Get pick for a specific week",
        "tags": [
          "Picks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entryId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the entry"
          },
          {
            "in": "path",
            "name": "entryNumber",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The entry number (1, 2, or 3)"
          },
          {
            "in": "path",
            "name": "week",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The NFL week number (1-18)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Pick"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "put": {
        "summary": "Update a pick",
        "tags": [
          "Picks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entryId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the entry"
          },
          {
            "in": "path",
            "name": "entryNumber",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The entry number (1, 2, or 3)"
          },
          {
            "in": "path",
            "name": "week",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The NFL week number (1-18)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "team": {
                    "type": "string",
                    "description": "The new team to pick"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Pick"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      },
      "delete": {
        "summary": "Delete a pick",
        "tags": [
          "Picks"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "entryId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the entry"
          },
          {
            "in": "path",
            "name": "entryNumber",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The entry number (1, 2, or 3)"
          },
          {
            "in": "path",
            "name": "week",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The NFL week number (1-18)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
  },
...
}